get_secret/requirements.txt
google.cloud==0.34.0
google-cloud-secret-manager==2.16.3
flask==2.3.3